Real mock data, transformation fix

+++ src/orchestrator/scan_data.py
@@ -7,4 +7,4 @@ from domain.dto.LaserScan import LaserScan
 
 @dataclass
 class ScanData:
-    lidar_data: Dict[float, List[LaserScan]] = field(default_factory=dict) # laser scan per step
+    lidar_data: Dict[int, List[LaserScan]] = field(default_factory=dict) # laser scan per step
